# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DACODES BACKEND",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://dacodesapi.test"
        }
    ],
    "paths": {
...
        "/api/admin/lessons": {
            "get": {
                "tags": [
                    "Admin Lessons"
                ],
                "summary": "Listar lecciones",
                "description": "Regresa una lista de lecciones",
                "operationId": "get_admin_lesson",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Lessons"
                ],
                "summary": "Alta de lección",
                "description": "Guardar una lección",
                "operationId": "save_admin_lesson",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "course_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/lessons/{lesson}": {
            "get": {
                "tags": [
                    "Admin Lessons"
                ],
                "summary": "Obtener detalle de una lección",
                "description": "Regresa un objeto con detalles de la lección",
                "operationId": "show_admin_lesson",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin Lessons"
                ],
                "summary": "Actualizar Lección",
                "description": "Actualizar valores de la lección",
                "operationId": "update_admin_lesson",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Lessons"
                ],
                "summary": "Eliminar lección",
                "description": "Eliminar una lección",
                "operationId": "delete_admin_lesson",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/questions": {
            "get": {
                "tags": [
                    "Admin Questions"
                ],
                "summary": "Listar Preguntas",
                "description": "Regresa una lista de preguntas",
                "operationId": "get_admin_question",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/questions/{id}": {
            "get": {
                "tags": [
                    "Admin Questions"
                ],
                "summary": "Obtener detalle de una pregunta",
                "description": "Regresa un objeto con detalles de una pregunta",
                "operationId": "show_admin_question",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/questions/{question}": {
            "delete": {
                "tags": [
                    "Admin Questions"
                ],
                "summary": "Eliminar pregunta",
                "description": "Eliminar una pregunta",
                "operationId": "delete_admin_question",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://dacodesapi.test/oauth/authorize",
                        "tokenUrl": "http://dacodesapi.test/oauth/token",
                        "refreshUrl": "http://dacodesapi.test/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}